robust calculator logic

+++ programming_paradigm/main.py
@@ -7,8 +7,8 @@ def main():
         sys.exit(1)
     else:
         try:
-            numerator = int(sys.argv[1])
-            denominator = int(sys.argv[2])
+            numerator = float(sys.argv[1])
+            denominator = float(sys.argv[2])
             result = safe_divide(numerator, denominator)
             print(result)
         except ValueError:
